+++ docs/changes.md
@@ -1,5 +1,8 @@
 # 最新动态
 
+* 2020/03/06
+  * 完善mutable image，支持自定义的图片创建函数。
+
 * 2020/03/05
   * 增加path\_replace\_extname函数。
   * 完善path\_replace\_basename函数。